Separates the getUserFoodFromBank and removeFoodFromUser functions Adds back the 1 > 2 to disable the food usage on groupkill customDenier
