+++ src/lib/minions/functions/removeFoodFromUser.ts
@@ -1,36 +1,10 @@
 import { KlasaClient, KlasaUser } from 'klasa';
 import { addBanks, removeBankFromBank } from 'oldschooljs/dist/util';
-import { O } from 'ts-toolbelt';
 
 import { Eatables } from '../../eatables';
 import { ClientSettings } from '../../settings/types/ClientSettings';
 import { UserSettings } from '../../settings/types/UserSettings';
-import { Bank, ItemBank } from '../../types';
-
-export function getUserFoodFromBank(
-	userBank: O.Readonly<Bank>,
-	totalHealingNeeded: number
-): false | ItemBank {
-	let totalHealingCalc = totalHealingNeeded;
-	let foodToRemove: Bank = {};
-	// Gets all the eatables in the user bank
-	for (const eatable of Eatables.sort((i, j) => (i.healAmount > j.healAmount ? 1 : -1))) {
-		const inBank = userBank[eatable.id];
-		const toRemove = Math.ceil(totalHealingCalc / eatable.healAmount);
-		if (!inBank) continue;
-		if (inBank >= toRemove) {
-			totalHealingCalc -= Math.ceil(eatable.healAmount * toRemove);
-			foodToRemove = addBanks([foodToRemove, { [eatable.id]: toRemove }]);
-			break;
-		} else {
-			totalHealingCalc -= Math.ceil(eatable.healAmount * inBank);
-			foodToRemove = addBanks([foodToRemove, { [eatable.id]: inBank }]);
-		}
-	}
-	// Check if qty is still above 0. If it is, it means the user doesn't have enough food.
-	if (totalHealingCalc > 0) return false;
-	return foodToRemove;
-}
+import getUserFoodFromBank from './getUserFoodFromBank';
 
 export default async function removeFoodFromUser(
 	client: KlasaClient,
